requirements.txt
discord.py
psutil
jishaku
youtube-dl
pynacl
FFmpeg
asyncio
discord-pretty-help